backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==3.0.0
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

